app/requirements.txt
flask
yfinance
cassandra-driver
